Update SPI message processing
Combine all of the SPI transfers in a message into a single transfer
so that one of the CP2130 primitive operations can be used.  This
allows for chip selection during the whole SPI message.  Also, convert
Read and WriteRead to issue asynchronous read request before the write
blocking write request.  This is in line with how the userspace
reference implementation from Silicon Labs implements the operations.

Signed-off-by: David Frey <dfrey@sierrawireless.com>

Conflicts:
	spi-cp2130.c